2017 Version
This is the original version of my augmented reality app The Big Sky Above Us, available in the app store. This version was built before the creation of AugmaOS and features unique code to render the solar system with realtime interaction and music! All code, music, assets etc are provided as is and may no longer be updated. This repository is to serve as an archive of the original software. Feel free to use everything inside as you please, I hope you learn something from it!